moved queries to their own file
